test_4/airflow-docker/requirements.txt
google-auth
google-analytics-data
google-cloud-bigquery
apache-airflow
pandas
numpy